+++ src/main/java/entities/SaveFileFactory.java
@@ -2,13 +2,13 @@ package entities;
 
 public class SaveFileFactory {
 
-    public SaveFile create(double timeLeft,
+    public SaveFile create(int timeLeft,
                            String name,
                            String spritePath,
-                           double hunger,
-                           double thirst,
-                           double cleanliness,
-                           double happiness) {
+                           int hunger,
+                           int thirst,
+                           int cleanliness,
+                           int happiness) {
         return new SaveFile(timeLeft, name, spritePath, hunger, thirst, cleanliness, happiness);
     }
 }
